Task4/requirements.txt
Flask==1.1.4
psycopg2-binary==2.9.1